requirements.txt
selenium~=4.10.0
driver~=0.1.0
python-dotenv~=1.0.0
psutil==5.9.5
structlog~=23.2.0
pytest~=7.4.3
pytest-xdist~=3.5.0